==> src/p01-install.R <==
# Installs the required R packages.
#
# version: 1.1
# Author: Vincent Labatut 06/2013
#
# setwd("~/eclipse/workspaces/Networks/Orleans/")
# setwd("C:/Eclipse/workspaces/Networks/Orleans/")
# source("src/p01-install.R")
###############################################################################


# power law tests
install.packages("igraph")


# Davies-Bouldin measure
install.packages("clusterSim")


# anova (unbalanced)
install.packages("car")
